# Use HashMap::entry api to speed up lookups