Currently in the hot loop for diffusion we're doing

```rust
        let existing = self.get(tile_pos);
        self.map.insert(tile_pos, existing + signal_strength);
```

Which involves two hash map lookups. We can use the entry api to turn this into a single hash map lookup with significant performance benefits (17-20%). Arguably the code is also cleaner this way.

Also added `AddAssign` and `SubAssign` implementations, just because it makes the code look a bit nicer.

Benchmarks

```
test result: ok. 0 passed; 0 failed; 51 ignored; 0 measured; 0 filtered out; finished in 0.00s

     Running benches/signals.rs (target/release/deps/signals-e82e00caa29fb839)
add_signal_minimal      time:   [212.55 ns 213.71 ns 214.64 ns]
                        change: [-1.2760% -0.5628% +0.1650%] (p = 0.13 > 0.05)
                        No change in performance detected.

signal_diffusion_minimal
                        time:   [612.12 ns 613.33 ns 614.46 ns]
                        change: [-5.0854% -4.8589% -4.6528%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 17 outliers among 100 measurements (17.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
  14 (14.00%) high severe

add_signal_tiny         time:   [13.901 µs 13.914 µs 13.933 µs]
                        change: [-2.3908% -2.2498% -2.1225%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  5 (5.00%) high mild
  2 (2.00%) high severe

signal_diffusion_tiny   time:   [72.077 µs 72.131 µs 72.199 µs]
                        change: [-17.871% -17.765% -17.641%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  2 (2.00%) high mild
  5 (5.00%) high severe

add_signal_modest       time:   [11.697 ms 11.711 ms 11.727 ms]
                        change: [-3.2473% -3.1040% -2.9383%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe

Benchmarking signal_diffusion_modest: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 7.1s, or reduce sample count to 70.
signal_diffusion_modest time:   [70.494 ms 70.542 ms 70.592 ms]
                        change: [-20.130% -19.996% -19.877%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
  2 (2.00%) high severe
```